refactor: enum package 변경
