+++ src/main/java/step2/domain/lotto/LottoNumberAutoGenerationStrategy.java
@@ -1,7 +1,5 @@
 package step2.domain.lotto;
 
-import step2.vo.LottoNumberValue;
-
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
