+++ src/test/java/step2/domain/statistics/MatchTest.java
@@ -2,7 +2,6 @@ package step2.domain.statistics;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import step2.vo.Rank;
 
 import java.util.HashMap;
 import java.util.Map;
