+++ src/test/java/step2/domain/statistics/ProfitTest.java
@@ -3,7 +3,6 @@ package step2.domain.statistics;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import step2.utils.ProfitTestUtil;
-import step2.vo.Rank;
 
 import static org.assertj.core.api.Assertions.assertThat;
