+++ src/test/java/step2/domain/statistics/StatisticsTest.java
@@ -4,7 +4,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import step2.domain.lotto.*;
 import step2.utils.ProfitTestUtil;
-import step2.vo.Rank;
 
 import java.util.Arrays;
